Refresh empty OAuth2 credentials in batch request
Batch request serialization requires the credentials in the http
object to be applied to every individual request *prior* to
actually invoking http.request(). This circumvents the self-refresh
logic in the request() function monkey-patched by the OAuth2
credentials, thus resulting in the use of an uninitialized credentials
object.

The fix here is to detect when OAuth2-like credentials are being
used and to force a refresh in case the access token is not
initialized yet.

Resolves: #211